# openapi.yaml
openapi: 3.1.0
info:
  title: Codestra WhatsApp Application API
  version: 0.2.0
  description: >-
    Business/application control plane for WhatsApp. Provider effects are submitted only through
    Middleware V3. W3 owns durable inbound events, conversations, routing, automation state,
    human handoff, audit readback and dead-letter recovery.
servers:
  - url: /
...
paths:
...
  /internal/v1/inbound-events:
    post:
      tags: [W3 Internal]
      security: [{ internalToken: [] }]
      summary: Durably accept one normalized inbound event
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InboundEvent" }
      responses:
        "200": { description: Duplicate event already accepted, content-equivalent }
        "202": { description: Event durably accepted for processing }
        "409": { description: Conflicting duplicate provider event }
        "422": { $ref: "#/components/responses/Error" }
  /internal/v1/inbound-events/{eventId}:
    get:
      tags: [W3 Internal]
      security: [{ internalToken: [] }]
      parameters:
        - { $ref: "#/components/parameters/EventId" }
      responses:
        "200": { description: Event processing readback }
        "404": { $ref: "#/components/responses/Error" }
...
  /platform/v1/whatsapp/campaigns/validate:
    post:
      tags: [W2]
      responses:
        "200": { description: Campaign valid }
        "422": { description: Campaign invalid }
  /platform/v1/whatsapp/messages:
    post:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      description: Submit an eligible WhatsApp effect to Middleware V3. Tenant/actor are derived from authenticated request context.
      parameters:
        - { name: X-Tenant-ID, in: header, required: true, schema: { type: string } }
        - { name: X-Command-ID, in: header, required: false, schema: { type: string } }
        - { name: X-Correlation-ID, in: header, required: false, schema: { type: string } }
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8 } }
      responses:
        "202": { description: Middleware accepted }
        "400": { description: Invalid request }
        "401": { description: Authentication required }
        "403": { description: Recipient, tenant, or role denied }
        "409": { description: Header/body correlation or idempotency mismatch }
        "415": { description: JSON content type required }
        "423": { description: Production/external sends disabled }
        "503": { description: Middleware WhatsApp command family not registered/configured }
  /platform/v1/whatsapp/conversations:
    get:
      tags: [W3 Conversations]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: assigned_to, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200": { description: Tenant-scoped paginated conversation list }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /platform/v1/whatsapp/conversations/{conversationId}:
    get:
      tags: [W3 Conversations]
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ConversationId" }
      responses:
        "200": { description: Conversation readback }
        "404": { $ref: "#/components/responses/Error" }
...
  /platform/v1/whatsapp/dead-letters:
    get:
      tags: [W3 Conversations]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Admin-only tenant-scoped dead-letter list }
...
  /platform/v1/whatsapp/operations/{operation_id}:
    get:
      security: [{ bearerAuth: [] }]
      description: Read the authoritative Middleware V3 operation through the WhatsApp application boundary.
      parameters:
        - in: path
          name: operation_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Middleware V3 operation status }
        "401": { description: Authentication required }
        "403": { description: Tenant or role denied }
        "404": { description: Operation not found }
  /platform/v1/whatsapp/me:
    get:
      tags: [W3 Conversations]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current tenant-scoped operator identity and roles }
  /platform/v1/whatsapp/dashboard:
    get:
      tags: [W3 Conversations]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant-scoped agent-console business and conversation summary }
  /platform/v1/whatsapp/contacts:
    get:
      tags: [W1]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant-scoped contact list }
    post:
      tags: [W1]
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Durable contact created }
        "422": { description: Contact validation failed }
  /platform/v1/whatsapp/contacts/{contactId}:
    get:
      tags: [W1]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: contactId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Contact readback }
        "404": { description: Contact not found }
    patch:
      tags: [W1]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: contactId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Contact updated with optimistic version check }
        "409": { description: Stale contact version }
  /platform/v1/whatsapp/templates:
    get:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant-scoped template list }
    post:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Durable local template draft created }
  /platform/v1/whatsapp/templates/{templateId}:
    get:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Template readback }
        "404": { description: Template not found }
    patch:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: templateId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Template updated with optimistic version check }
        "409": { description: Stale template version }
  /platform/v1/whatsapp/campaigns:
    get:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Tenant-scoped campaign list }
    post:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Durable campaign draft created }
        "422": { description: Campaign cannot enter ready state until validation passes }
  /platform/v1/whatsapp/campaigns/{campaignId}:
    get:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Campaign readback }
        "404": { description: Campaign not found }
    patch:
      tags: [W2]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: campaignId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Campaign updated with optimistic version check }
        "409": { description: Stale campaign version }
        "422": { description: Campaign validation failed }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT with subject, tenant and whatsapp_agent/supervisor/admin role.
    internalToken:
      type: apiKey
      in: header
      name: x-internal-token
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string }
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    InboundEvent:
      type: object
      additionalProperties: false
      required: [schema_version, event_id, provider_event_id, correlation_id, tenant_id, channel, sender_identity, recipient_identity, event_type, provider_timestamp, ingested_at, payload_hash, content, metadata]
      properties:
        schema_version: { const: whatsapp.inbound.v1 }
        event_id: { type: string, format: uuid }
        provider_event_id: { type: string, minLength: 1 }
        correlation_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        channel: { const: whatsapp }
        sender_identity: { type: string, minLength: 1 }
        recipient_identity: { type: string, minLength: 1 }
        event_type: { type: string, enum: [message.received, message.status, conversation.updated, contact.updated] }
        provider_timestamp: { type: string, format: date-time }
        ingested_at: { type: string, format: date-time }
        payload_hash: { type: string, pattern: "^[a-f0-9]{64}$" }
        content:
          type: object
          required: [type]
          additionalProperties: true
          properties:
            type: { type: string }
            text: { type: [string, "null"] }
        metadata: { type: object, additionalProperties: true }
...
  responses:
    Error:
      description: Fail-closed error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message, retryable]
                properties:
                  code: { type: string }
                  message: { type: string }
                  retryable: { type: boolean }